+++ service/devicefarm/examples_test.go
@@ -1182,11 +1182,11 @@ func ExampleDeviceFarm_ListSamples_shared00() {
 // To get information about suites
 //
 // The following example returns information about suites, given a specific Device Farm
-// project.
+// job.
 func ExampleDeviceFarm_ListSuites_shared00() {
 	svc := devicefarm.New(session.New())
 	input := &devicefarm.ListSuitesInput{
-		Arn:       aws.String("arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456"),
+		Arn:       aws.String("arn:aws:devicefarm:us-west-2:123456789101:job:EXAMPLE-GUID-123-456"),
 		NextToken: aws.String("RW5DdDJkMWYwZjM2MzM2VHVpOHJIUXlDUXlhc2QzRGViYnc9SEXAMPLE"),
 	}